Finished styleze

+++ src/core/lmdb/lmdb_config.cpp
@@ -18,6 +18,9 @@
 
 #include "core/lmdb/lmdb_config.h"
 
+#include <string>
+#include <vector>
+
 #include "common/sprintf.h"
 #include "common/file_system.h"
 
@@ -28,13 +31,13 @@ namespace {
 
 void parseOptions(int argc, char **argv, lmdbConfig& cfg) {
   for (int i = 0; i < argc; i++) {
-      int lastarg = i==argc-1;
+      int lastarg = i == argc-1;
 
-      if (!strcmp(argv[i],"-d") && !lastarg) {
+      if (!strcmp(argv[i], "-d") && !lastarg) {
           cfg.mb_delim_ = argv[++i];
       } else if (!strcmp(argv[i], "-f") && !lastarg) {
           cfg.dbname_ = argv[++i];
-      } else if (!strcmp(argv[i],"-c")) {
+      } else if (!strcmp(argv[i], "-c")) {
           cfg.create_if_missing_ = true;
       } else {
           if (argv[i][0] == '-') {
@@ -51,27 +54,27 @@ void parseOptions(int argc, char **argv, lmdbConfig& cfg) {
   }
 }
 
-}
+}  // namespace
 
 lmdbConfig::lmdbConfig()
- : LocalConfig(common::file_system::prepare_path("~/test.lmdb")), create_if_missing_(false) {
+  : LocalConfig(common::file_system::prepare_path("~/test.lmdb")), create_if_missing_(false) {
 }
 
-}
+}  // namespace fastonosql
 
 namespace common {
 
 std::string convertToString(const fastonosql::lmdbConfig &conf) {
   std::vector<std::string> argv = conf.args();
 
-  if(conf.create_if_missing_){
+  if (conf.create_if_missing_) {
       argv.push_back("-c");
   }
 
   std::string result;
-  for(int i = 0; i < argv.size(); ++i){
+  for (size_t i = 0; i < argv.size(); ++i) {
       result += argv[i];
-      if(i != argv.size()-1){
+      if (i != argv.size()-1) {
           result += " ";
       }
   }
@@ -87,7 +90,7 @@ fastonosql::lmdbConfig convertFromString(const std::string& line) {
   char *argv[kMaxArgs] = {0};
 
   char* p2 = strtok((char*)line.c_str(), " ");
-  while(p2){
+  while (p2) {
       argv[argc++] = p2;
       p2 = strtok(0, " ");
   }
@@ -96,4 +99,4 @@ fastonosql::lmdbConfig convertFromString(const std::string& line) {
   return cfg;
 }
 
-}
+}  // namespace comon
